Issue #5272 Make o.e.j.security.User class inner to UserStore
