+++ jetty-security/src/main/java/org/eclipse/jetty/security/UserStore.java
@@ -18,9 +18,12 @@
 
 package org.eclipse.jetty.security;
 
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 
 import org.eclipse.jetty.util.component.AbstractLifeCycle;
 import org.eclipse.jetty.util.security.Credential;
@@ -32,6 +35,32 @@ import org.eclipse.jetty.util.security.Credential;
 public class UserStore extends AbstractLifeCycle
 {
     protected final Map<String, User> _users = new ConcurrentHashMap<>();
+    
+    protected class User
+    {
+        protected UserPrincipal _userPrincipal;
+        protected List<RolePrincipal> _rolePrincipals = Collections.emptyList();
+        
+        protected User(String username, Credential credential, String[] roles)
+        {
+            _userPrincipal = new UserPrincipal(username, credential);
+
+            _rolePrincipals = Collections.emptyList();
+            
+            if (roles != null)
+                _rolePrincipals = Arrays.stream(roles).map(RolePrincipal::new).collect(Collectors.toList());
+        }
+        
+        protected UserPrincipal getUserPrincipal()
+        {
+            return _userPrincipal;
+        }
+        
+        protected List<RolePrincipal> getRolePrincipals()
+        {
+            return _rolePrincipals;
+        }
+    }
     
     public void addUser(String username, Credential credential, String[] roles)
     {
